week 2 adjustments
